+++ styleguides/admin/admin-ui/src/components/Menu/components/ActionButton.tsx
@@ -1,8 +1,8 @@
 import React, { Ref } from 'react'
 import { forwardRef } from '@vtex/admin-ui-system'
 
-import { IconActions, IconCaret } from '../../icons'
-import { Button, ButtonProps } from '../Button'
+import { IconActions, IconCaret } from '../../../icons'
+import { Button, ButtonProps } from '../../Button'
 
 /**
  * Button used inside of menu's disclosure