+++ styleguides/admin/admin-ui/src/components/Menu/components/AriaMenu.tsx
@@ -1,0 +1,19 @@
+import {
+  useMenuState,
+  Menu as ReakitMenu,
+  MenuItem as ReakitMenuItem,
+  MenuButton as ReakitMenuButton,
+  MenuState,
+  MenuStateReturn,
+  MenuSeparator as ReakitMenuSeparator,
+} from 'reakit/Menu'
+
+export {
+  ReakitMenuButton,
+  ReakitMenuSeparator,
+  ReakitMenu,
+  ReakitMenuItem,
+  MenuStateReturn,
+  MenuState,
+  useMenuState,
+}