+++ styleguides/admin/admin-ui/src/components/Menu/index.ts
@@ -1,10 +1,3 @@
-export { Menu, MenuProps } from './Stateful'
-export {
-  StatelessMenu,
-  useMenuState,
-  StatelessMenuProps,
-  MenuItemProps,
-  MenuStateReturn,
-  MenuDisclosure,
-} from './Stateless'
-export { ActionButton, ActionButtonProps } from './ActionButton'
+export * from './Stateful'
+export * from './Stateless'
+export * from './components'