+++ styleguides/brand-ui/src/index.ts
@@ -34,3 +34,10 @@ export {
   Timeline,
   EventProps as TimelineEventProps,
 } from './components/Timeline'
+export {
+  Calendar,
+  CalendarProps,
+  Events,
+  DayProps,
+} from './components/Calendar'
+export { Carousel, CarouselProps } from './components/Carousel'