+++ styleguides/brand-ui/src/theme/theme.ts
@@ -10,6 +10,8 @@ import popover from './components/popover'
 import helloBar from './components/helloBar'
 import localeSwitcher from './components/localeSwitcher'
 import timeline from './components/timeline'
+import calendar from './components/calendar'
+import carousel from './components/carousel'
 
 export const theme = {
   colors: {
@@ -72,6 +74,8 @@ export const theme = {
   helloBar,
   localeSwitcher,
   timeline,
+  calendar,
+  carousel,
   fonts: {
     body: 'sans-serif',
     heading: 'sans-serif',